Install 3D headers
