+++ src/3d/mesh/qgsmeshterraingenerator.h
@@ -23,32 +23,8 @@
 #include "qgsmeshlayer.h"
 #include "qgstriangularmesh.h"
 #include "qgsterraingenerator.h"
-#include "qgsterraintileloader_p.h"
 
 #define SIP_NO_FILE
-
-///@cond PRIVATE
-
-//! Chunk loader for mesh terrain implementation
-class QgsMeshTerrainTileLoader: public QgsTerrainTileLoader
-{
-    Q_OBJECT
-  public:
-    //! Construct the loader for a node
-    QgsMeshTerrainTileLoader( QgsTerrainEntity *terrain,
-                              QgsChunkNode *node,
-                              const QgsTriangularMesh &triangularMesh,
-                              const QgsMesh3DSymbol *symbol );
-
-    //! Create the 3D entity and returns it
-    Qt3DCore::QEntity *createEntity( Qt3DCore::QEntity *parent ) override;
-
-  private:
-    QgsTriangularMesh mTriangularMesh;
-    std::unique_ptr< QgsMesh3DSymbol > mSymbol;
-};
-
-///@endcond
 
 /**
  * \ingroup 3d
