+++ src/3d/qgs3dsceneexporter.cpp
@@ -64,8 +64,8 @@ typedef Qt3DCore::QGeometry Qt3DQGeometry;
 
 #include "qgsbillboardgeometry.h"
 #include "qgsterraintileentity_p.h"
-#include "qgsterrainentity_p.h"
-#include "qgschunknode_p.h"
+#include "qgsterrainentity.h"
+#include "qgschunknode.h"
 #include "qgsterraingenerator.h"
 #include "qgs3dmapsettings.h"
 #include "qgsflatterraingenerator.h"
@@ -76,6 +76,7 @@ typedef Qt3DCore::QGeometry Qt3DQGeometry;
 #include "qgsterraintextureimage_p.h"
 #include "qgsterraintexturegenerator_p.h"
 #include "qgsmeshterraingenerator.h"
+#include "qgsmeshterraintileloader_p.h"
 #include "qgsvectorlayer.h"
 #include "qgsabstract3drenderer.h"
 #include "qgsvectorlayer3drenderer.h"
