+++ src/3d/qgs3dutils.cpp
@@ -29,8 +29,8 @@
 #include "qgsabstract3dengine.h"
 #include "qgsterraingenerator.h"
 #include "qgscameracontroller.h"
-#include "qgschunkedentity_p.h"
-#include "qgsterrainentity_p.h"
+#include "qgschunkedentity.h"
+#include "qgsterrainentity.h"
 #include "qgsraycastingutils_p.h"
 #include "qgspointcloudrenderer.h"
 #include "qgspointcloud3dsymbol.h"
