+++ src/3d/terrain/qgsdemterraintileloader_p.cpp
@@ -16,12 +16,12 @@
 #include "qgsdemterraintileloader_p.h"
 
 #include "qgs3dmapsettings.h"
-#include "qgschunknode_p.h"
+#include "qgschunknode.h"
 #include "qgsdemterraingenerator.h"
 #include "qgsdemterraintilegeometry_p.h"
 #include "qgseventtracing.h"
 #include "qgsonlineterraingenerator.h"
-#include "qgsterrainentity_p.h"
+#include "qgsterrainentity.h"
 #include "qgsterraintexturegenerator_p.h"
 #include "qgsterraintileentity_p.h"
 #include "qgsterraingenerator.h"
