+++ src/3d/terrain/qgsflatterraingenerator.h
@@ -19,7 +19,7 @@
 #include "qgis_3d.h"
 
 #include "qgsterraingenerator.h"
-#include "qgsterraintileloader_p.h"
+#include "qgsterraintileloader.h"
 #include <Qt3DExtras/QPlaneGeometry>
 
 #define SIP_NO_FILE
