+++ src/3d/terrain/qgsquantizedmeshterraingenerator.cpp
@@ -14,8 +14,8 @@
  ***************************************************************************/
 
 #include "qgsquantizedmeshterraingenerator.h"
-#include "qgschunkloader_p.h"
-#include "qgschunknode_p.h"
+#include "qgschunkloader.h"
+#include "qgschunknode.h"
 #include "qgscoordinatetransform.h"
 #include "qgslogger.h"
 #include "qgsmesh3dentity_p.h"
@@ -25,14 +25,14 @@
 #include "qgsquantizedmeshtiles.h"
 #include "qgsrectangle.h"
 #include "qgsterraintileentity_p.h"
-#include "qgsterraintileloader_p.h"
+#include "qgsterraintileloader.h"
 #include "qgstiledsceneindex.h"
 #include "qgstiledscenelayer.h"
 #include "qgstiledscenetile.h"
 #include "qgstiles.h"
 #include "qgstriangularmesh.h"
 #include "qgsgltf3dutils.h"
-#include "qgsterrainentity_p.h"
+#include "qgsterrainentity.h"
 #include "qgs3dmapsettings.h"
 #include "qgsvector3d.h"
 #include "qgsapplication.h"
